requirements.txt
requests
beautifulsoup4
urlextract
pandas
rich
python-dotenv
google-genai
